requirements.txt
ads1015==0.0.8
astral==3.2
autopep8==2.0.0
certifi==2020.6.20
cffi==1.14.5
chardet==4.0.0
colorzero==1.1
distro==1.5.0
enviroplus==0.0.6
font-roboto==0.0.1
fonts==0.0.3
gpiozero==1.6.2
i2cdevice==0.0.7
idna==2.10
influxdb-client==1.34.0
ltr559==0.1.1
numpy==1.19.5
paho-mqtt==1.6.1
picamera2==0.3.5
pidng==4.0.9
piexif==1.1.3
Pillow==8.1.2
pimoroni-bme280==0.1.1
pkg_resources==0.0.0
ply==3.11
pms5003==0.0.5
psutil==5.9.4
pycodestyle==2.9.1
pycparser==2.20
pyserial==3.5
python-apt==2.2.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-prctl==1.7
pytz==2022.6
reactivex==4.0.4
requests==2.25.1
RPi.GPIO==0.7.0
simplejpeg==1.6.4
six==1.16.0
sounddevice==0.4.5
spidev==3.5
ssh-import-id==5.10
ST7735==0.0.4.post1
toml==0.10.1
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.5
v4l2-python3==0.3.1
